Refactor to remove promise hell
